Fixed problem with a train seeming to not respect the switch of a switched join segment after being dragged to the intersection of the two switched paths.
A little more detail: A train dragged to the end point of a switched join segment would not respect the switch value.  Both paths would be calculated as "close", and the closer one would be chosen according to the search precision, which wouldn't respect the user's idea of which path the train should be on (according to the switch).