Fix crash on dragging a deletion from toolbar.
The deletions and exports toolbars were failing to update after adding
or removing a deletion or export.  This caused a problem in deletions
in particular, since a lingering deletion would try to let itself be
dragged out, but would trigger an exception when the segments file
could not be found.